+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Versions
 
+## x.x.x
+    * Add support for segment name.
+    * Add support for impression-level user revenue api.
+    * Add support for waterfall information api.
+    * Add support for custom data.
 ## 3.1.0
     * Add support for data passing.
     * Add API for Android to provide a built-in consent flow that sets the user consent flag.